services/files/requirements.txt
# Files Service Dependencies

# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# MinIO S3 Client
minio==7.2.0

# Redis
redis==5.0.1
aioredis==2.0.1

# File Processing
openpyxl==3.1.2
pandas==2.1.4
xlrd==2.0.1

# Antivirus
pyclamd==0.4.0

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
aiofiles==23.2.1

# File Type Detection
python-magic==0.4.27

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1